+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,28 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication"
         tools:targetApi="31">
+
         <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <!-- Add LoginActivity -->
+        <activity
+            android:name=".LoginActivity"
+            android:exported="true">
+        </activity>
+
+        <!-- Add HomeActivity -->
+        <activity
+            android:name=".HomeActivity"
+            android:exported="true">
+        </activity>
+
     </application>
 
 </manifest>
