+++ app/src/main/java/com/example/myapplication/HomeActivity.java
@@ -1,0 +1,37 @@
+package com.example.myapplication;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.ImageButton;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class HomeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        ImageButton btnRocket = findViewById(R.id.icon_rocket);
+        ImageButton btnProfile = findViewById(R.id.icon_Profile);
+        btnRocket.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                setContentView(R.layout.page_plan);
+
+            }
+        });
+        btnProfile.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                setContentView(R.layout.page_profile);
+
+            }
+        });
+    }
+}
+
+
