+++ app/src/main/java/com/example/myapplication/LoginActivity.java
@@ -1,5 +1,5 @@
 package com.example.myapplication;
-/*
+
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -13,23 +13,23 @@ public class LoginActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.page_login);
-        Button loginBtn = findViewById(R.id.btnLogin);
-        EditText usernameEditText = findViewById(R.id.username_EditText);
-        EditText passwordEditText = findViewById(R.id.password_EditText);
 
-        loginBtn.setOnClickListener(v -> {
-            if("admin".equals(usernameEditText.getText().toString()) &&
-               "admin".equals(passwordEditText.getText().toString())){
-                Intent intent = new Intent(LoginActivity.this, MainActivity.class);
-                startActivity(intent);
-            }
-        });
+
+            setContentView(R.layout.page_login);
+            Button loginBtn = findViewById(R.id.btnLogin);
+            EditText usernameEditText = findViewById(R.id.username_EditText);
+            EditText passwordEditText = findViewById(R.id.password_EditText);
+
+            loginBtn.setOnClickListener(v -> {
+                if ("admin".equals(usernameEditText.getText().toString()) &&
+                     "admin".equals(passwordEditText.getText().toString())) {
+                    UserManager.setUserLoggedIn(true);
+                    Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+                    startActivity(intent);
+                }
+            });
+
+
     }
-
-
-
-
 }
 
-*/
